requirements_txt.txt
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2

requests==2.31.0
urllib3==2.0.7

scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.3

spacy==3.6.1
nltk==3.8.1

pytesseract==0.3.10
pillow==10.0.0

python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0

pyyaml==6.0.1
toml==0.10.2

sqlite3

logging
smtplib
email
hashlib
json
os
sys
datetime
typing
re
urllib
pickle
random
time
argparse
threading
subprocess
collections

gunicorn==21.2.0

pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

black==23.9.1
flake8==6.1.0
mypy==1.6.1